requirements.txt
tkinter
arabic_reshaper
python-bidi
pyperclip